backend/requirements.txt
flask
python-dotenv
requests
pytesseract
pdfminer.six
werkzeug
gunicorn